cmd/jujud: move the unit agent out of the main package
The unit agent code now lives in the cmd/jujud/agent package so that
it can be imported for testing elsewhere (e.g. the feature
tests). This also allows a lot of duplicated test code to be removed.
